fix: reset new password validator

+++ app/Http/Controllers/Frontend/ResetPasswordController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers\Frontend;
 
+use App\Commons\Message;
 use App\Commons\Response;
 use Illuminate\Http\Request;
 use App\Http\Controllers\Controller;
@@ -9,6 +10,7 @@ use App\Models\User;
 use App\Models\PasswordReset;
 use App\Notifications\ResetPasswordRequest;
 use Carbon\Carbon;
+use Illuminate\Support\Facades\Validator;
 use Illuminate\Support\Str;
 
 
@@ -39,7 +41,7 @@ class ResetPasswordController extends Controller
             $user->notify(new ResetPasswordRequest($passwordReset->token));
         }
 
-        return $this->response->formatResponse(200, [], 'Password reset email has been sent.');
+        return $this->response->formatResponse(200, [], 'Password reset email telah dikirim kan ke email kamu.');
     }
 
     public function reset(Request $request)
@@ -61,9 +63,28 @@ class ResetPasswordController extends Controller
             return back()->with('error', 'We can\'t find a user with that e-mail address.');
         }
 
-        $updatePasswordUser = $user->update(['password' => bcrypt($request->password)]);
-        $passwordReset->delete();
+        $rules = [
+            'password' => ['required', 'string', 'min:8', 'confirmed'],
+        ];
+        $message_errors = [
+            'required' => ':attribute wajib di isi,',
+            'confirmed'=>':attribute harus sama dengan confirm :attribute,',
+            'min'=>':attribute minimal 8 karakter'];
+        $validator = Validator::make($request->all(), $rules, $message_errors);
 
-        return back()->with('success', 'Your password has been reset successfully. You can login with your new password now!');
+        if ($validator->fails()) {
+            $respValidator = $validator->messages();
+
+            return back()->with('error', Message::genErrorMessage($respValidator)) ;
+           
+        } else {
+            $updatePasswordUser = $user->update(['password' => bcrypt($request->password)]);
+            $passwordReset->delete();
+            return back()->with('success', 'Your password has been reset successfully. You can login with your new password now!');
+        }
+
+        
+
     }
+    
 }
